Implement PC sound driver wrappers
